// 01-app-typescript/src/interfaces/reqResp.interface.tsx
// Generated by https://quicktype.io

export interface ReqRespListado {
    page:        number;
    per_page:    number;
    total:       number;
    total_pages: number;
    data:        UsuariosList[];
    support:     Support;
}

export interface UsuariosList {
    id:         number;
    email:      string;
    first_name: string;
    last_name:  string;
    avatar:     string;
}

export interface Support {
    url:  string;
    text: string;
}